refactor & upsert
